add: new feat

+++ Dookki_Web/Models/Cart.cs
@@ -10,14 +10,14 @@ namespace Dookki_Web.Models
         DOOKKIEntities db = new DOOKKIEntities();
         public int ticketID { get; set; }
         public string name { get; set; }
-        public float price { get; set; }
+        public decimal price { get; set; }
         public int quantily { get; set; }
         public Cart(int ticketID) 
         {
             this.ticketID = ticketID;
             var ticket  = db.Tickets.Single(n=>n.ID == ticketID);
             name = ticket.Name;
-            price = (int)ticket.Price;
+            price = ticket.Price;
             quantily = 1;
         }
     }
